Moved extension methods for rendering sprite into `SpriteBatch.Extensions`
