+++ source/MonoGame.Extended/Graphics/SpriteBatch.Extensions.cs
@@ -44,6 +44,35 @@ public static class SpriteBatchExtensions
 
     #endregion -------------------------NinePatch-----------------------------
 
+    #region ----------------------------Sprite-----------------------------
+    public static void Draw(this Sprite sprite, SpriteBatch spriteBatch, Vector2 position, float rotation, Vector2 scale)
+    {
+        Draw(spriteBatch, sprite, position, rotation, scale);
+    }
+
+    public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Transform2 transform)
+    {
+        Draw(spriteBatch, sprite, transform.Position, transform.Rotation, transform.Scale);
+    }
+
+    public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Vector2 position, float rotation = 0)
+    {
+        Draw(spriteBatch, sprite, position, rotation, Vector2.One);
+    }
+
+    public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Vector2 position, float rotation, Vector2 scale)
+    {
+        if (sprite == null) throw new ArgumentNullException(nameof(sprite));
+
+        if (sprite.IsVisible)
+        {
+            var texture = sprite.TextureRegion.Texture;
+            var sourceRectangle = sprite.TextureRegion.Bounds;
+            spriteBatch.Draw(texture, position, sourceRectangle, sprite.Color * sprite.Alpha, rotation, sprite.Origin, scale, sprite.Effect, sprite.Depth);
+        }
+    }
+    #endregion -------------------------Sprite-----------------------------
+
     #region ----------------------------Texture2D-----------------------------
 
     public static void Draw(this SpriteBatch spriteBatch, Texture2D texture, Rectangle sourceRectangle, Rectangle destinationRectangle, Color color, Rectangle? clippingRectangle)
